Remove persistence mechanism when send a message
When a message was sent a queue was declared to receive that message
and keep it. This was done to prevent the message from being lost if
BabelTower was not yet connected. However, VirtualThing would need to
know the BabelTower queue information, which is not recommended.
This patch removes that mechanism.